refactor(tracing): split into package (config, providers, spans, context) with re-exports
Replace monolithic tracing module with modular package. Preserve imports and API; tests updated to import from package surface continue to pass.
