+++ src/meridian/observability/tracing/__init__.py
@@ -1,0 +1,29 @@
+from __future__ import annotations
+
+from .config import TracingConfig
+from .providers import (
+    get_tracer,
+    configure_tracing,
+    is_tracing_enabled,
+    InMemoryTracer,
+    NoopTracer,
+)
+from .spans import Span, NoopSpan, OpenTelemetrySpan
+from .context import start_span, set_trace_id, get_trace_id, get_span_id, generate_trace_id
+
+__all__ = [
+    "TracingConfig",
+    "get_tracer",
+    "configure_tracing",
+    "is_tracing_enabled",
+    "InMemoryTracer",
+    "NoopTracer",
+    "Span",
+    "NoopSpan",
+    "OpenTelemetrySpan",
+    "start_span",
+    "set_trace_id",
+    "get_trace_id",
+    "get_span_id",
+    "generate_trace_id",
+]
